Simplify and fix cis2 auth mock usage
While fixing the feature specs for the change to the order that we check
authentication details, it was obvious that the way mock_cis2_auth works
is much too convoluted. This fixes that aspect of the test but also
allows us to greatly simplify the interface to mock_cis2_auth in a
future commit.